% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{bmod1}
\alias{bmod1}
\title{Fits CES boundary model 1}
\usage{
bmod1(data, f, nest)
}
\arguments{
\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}

\item{y}{the response variable, assumed to be a vector}

\item{x1}{the first factor of production, assumed to be a vector}

\item{time}{the indexed time variable, assumed to be a vector}

\item{formula}{the original CES formula that is being fitted}
}
\value{
an \code{lm} model object with \code{naturalCoeffs} and \code{meta} attributes.
}
\description{
Fits CES boundary model 1
}

