% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{bmod2}
\alias{bmod2}
\title{Fits CES boundary model 2}
\usage{
bmod2(data, f, nest)
}
\arguments{
\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}

\item{y}{the response variable, assumed to be a vector}

\item{x2}{the second factor of production, assumed to be a vector}

\item{time}{the indexed time variable, assumed to be a vector}

\item{formula}{the original CES formula that is being fitted}
}
\value{
an \code{lm} model object with \code{naturalCoeffs} and \code{meta} attributes.
}
\description{
The boundary model is \code{y = gamma * A * x2} in log transform space
as {ln(y/x2) = ln(gamma) + lambda*time}.
For ces boundary model 2, the constraints are \code{delta_1 = 0}, and \code{delta = 1}.
\code{rho_1} (\code{sigma_1}) and \code{rho} (\code{sigma}) are unknowable and set to \code{NA}.
}

