% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{bmod3}
\alias{bmod3}
\title{Fits CES boundary model 3}
\usage{
bmod3(y, x1, x2, time, formula, nest)
}
\arguments{
\item{y}{the response variable, assumed to be a vector}

\item{time}{the indexed time variable, assumed to be a vector}

\item{formula}{the original CES formula that is being fitted}

\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}

\item{minx1x2}{the factor of production, assumed to be a vector}
}
\description{
The boundary model is \code{y = gamma * A * min(x1, x2)} in log transform space
as {ln(y/min(x1, x2)) = ln(gamma) + lambda*time}.
For ces boundary model 3, the constraints are \code{sigma_1 = 0}, and \code{delta = 1}.
\code{delta_1} and \code{rho} (\code{sigma}) are unknowable and set to \code{NA}.
}

