% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{cd2Model}
\alias{cd2Model}
\title{Fitting Cobb-Douglas Models}
\usage{
cd2Model(formula, data, response, x1, x2, time, constrained = FALSE,
  save.data = TRUE, ...)
}
\arguments{
\item{data}{a data fram in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{constrained}{a logical indicating whether the parameters are constrained}

\item{formala}{a formual of the form \code{response ~ x1 + x2 + time}}
}
\value{
a CDEmodel object, which is an lm object with some additional attributes.
}
\description{
Fitting Cobb-Douglas Models
}
\examples{
US <- subset(Calvin, Country=="US")
cd2Model(response = iGDP, x1 = iK, x2 = iL, time = iYear, data=US)
cd2Model(iGDP ~ iK + iL + iYear, data=US)
}

