% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cdeModel}
\alias{cdeModel}
\title{Fitting Cobb-Douglas models with Energy}
\usage{
cdeModel(formula, data, response, capital, labor, energy, time,
  constrained = FALSE, save.data = TRUE, ...)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ capital + labor + energy + time}}

\item{data}{a data frame in which \code{formala} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{capital}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{labor}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{energy}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{constrained}{a logical indicated whether the coefficents are constrained. See details}

\item{\dots}{additional arguments; currently unused.}
}
\description{
Fitting Cobb-Douglas models with Energy
}
\details{
More about contranints TBA.
}

