% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cdModel}
\alias{cdModel}
\title{Fitting Cobb-Douglas models}
\usage{
cdModel(formula, data, response, capital, labor, energy, time,
  constrained = FALSE, save.data = TRUE, ...)
}
\arguments{
\item{data}{a data frame in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{capital}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{labor}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{energy}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{constrained}{a logical indicating whether the parameters are constrained}

\item{formala}{a formula of the form
\code{response ~ capital + labor + time} or
\code{response ~ capital + labor + energy + time}}
}
\value{
a CDEmodel object, which is an lm object with some additioanl attributes.
}
\description{
Fitting Cobb-Douglas models
}

