% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{cesBoundaryModel}
\alias{cesBoundaryModel}
\title{Fits CES boundary models}
\usage{
cesBoundaryModel(formula, data, nest, id)
}
\arguments{
\item{formula}{a CES formula in the form \code{y ~ a + b + c + d + time}}

\item{data}{historical time series data}

\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}

\item{id}{the identification number (from Table 2) for the boundary model you want to fit.}
}
\value{
a model object with class \code{CESModel} and \code{naturalCoeffs} and \code{meta} attributes.
}
\description{
The boundary models are given in Table 2 of
Heun, et al "An Empirical Investigation of the Role of Energy in Economic Growth"
}
\note{
the \code{naturalCoeffs} attribute includes correct values of boundary parameters.
\code{NA} values in \code{naturalCoeffs} indicate that the parameter is unknowable at that boundary.
}
\examples{
if (require(EconData) & require(dplyr)) {
  cesBoundaryModel(iGDP ~ iK + iL + iQp + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(1,2,3), id=1)
  cesBoundaryModel(iGDP ~ iK + iL + iQp + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(1,2,3), id=2)
  cesBoundaryModel(iGDP ~ iK + iL + iQp + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(1,2,3), id=3)
  }
}

