% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesModel.R
\name{cesModel}
\alias{cesModel}
\title{Fitting CES models}
\usage{
cesModel(formula, data, response, x1 = NULL, x2 = NULL, x3 = NULL,
  x4 = NULL, time, nest = 1:3, prevModel = NULL, algorithms = c("PORT",
  "L-BFGS-B"), multErr = TRUE, rho = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5,
  -0.75, -0.9, -0.99), rho1 = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5, -0.75,
  -0.9, -0.99), digits = 6, save.data = TRUE, constrained = TRUE,
  fitBoundaries = TRUE, ...)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ a + b + c + d + time}.
\code{c} and \code{d} are optional.}

\item{data}{a data frame, in which to evaluate the formula.}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x3}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x4}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{nest}{a permutation (a,b,c,d) of the integers 1 through 4.
For models with 3 factors, the nesting
is (xa + xb) + xc.  For 4 factors, the nesting is (xa + xb) + (xc + xd)}

\item{prevModel}{a model used to start a gradient search.
\code{prevModel} will be used as a starting point for a gradient search after
(a) grid search in \code{rho} and \code{rho1} and
(b) a gradient search starting from the best grid search
are complete.
#' Use \code{NULL} (the default value)
if you want to skip the gradient search from a previous model.}

\item{multErr}{logical, tells whether errors are to be assessed in a multiplicative manner}

\item{rho,rho1}{Default values for \code{rho} and \code{rho1} are a grid upon which
searches will be made.
Note that \code{rho = 0.25} and \code{rho1 = 0.25} are included. These are the default
starting values for \code{rho} and \code{rho1}, so that we don't need to do a fit from
the default values.
\code{rho = 0.25} corresponds to \code{sigma = 0.8}.}

\item{digits}{the number of sse digits that is to be considered significant
when comparing one fit against another.}

\item{save.data}{a logical indicating whether data is to be saved with the model.
Be sure to set \code{TRUE} if resampling is needed later. Default is \code{TRUE}.}

\item{constrained}{a logical indicating whether the parameters should be constrained in the fitting process.
Default is \code{TRUE}.}

\item{fitBoundaries}{a logical indicating whether fits should be performed along boundaries.
Default is \code{TRUE}.}
}
\value{
a cesEst model with additional information attached as attributes.
}
\description{
This function fits a CES model
}
\note{
For now the components in \code{f} (or the arguments \code{response},
\code{x1}, \code{x2}, \code{x3}, \code{x4}, and \code{time}) must correspond to variables in \code{data} and may
not be other kinds of expressions.

For now, this function works for only 2 or 3 factors of production.
Setting the value of \code{x4} or using \code{f} of the form \code{y ~ a + b + c + d + time}
will not work.
}
\examples{
if (require("EconData") & require("dplyr")) {
  cesModel(iGDP ~ iK + iL + iQp + iYear, data = filter(Calvin, Country=="US"), nest = c(1,2,3))
  cesModel(iGDP ~ iK + iL + iQp + iYear, data = filter(Calvin, Country=="US"), nest = c(2,3,1))
  cesModel(iGDP ~ iK + iL + iQp + iYear, data = filter(Calvin, Country=="US"), nest = c(3,1,2))
  cesModel(iGDP ~ iK + iL + iQp + iYear, data = filter(Calvin, Country=="ZM"), nest = c(3,1,2))
  cesModel(response="iGDP", x1="iK", x2="iL", x3="iQp", time="iYear",
           data = filter(Calvin, Country=="ZM"), nest = c(3,1,2))
}
}

