% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{cesTimeSeries}
\alias{cesTimeSeries}
\title{Extracts y, x1, x2, x3, x4, and time data as time series for CES boundary models}
\usage{
cesTimeSeries(f, data, nest)
}
\arguments{
\item{f}{the CES formula for which time series data is to be extracted,
assumed to be of the form \code{y ~ x1 + x2 + x3 + x4 + time}.}

\item{data}{the data frame from which time series data is to be extracted}

\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}
}
\value{
a named list of time series'.
Names are \code{y}, \code{x1}, \code{x2}, \code{x3}, \code{x4}, and \code{time}.
}
\description{
Given formula \code{f} and nest \code{nest}, this function
extracts a response variable (\code{y}),
factors of production (\code{x1}, \code{x2}, \code{x3}, and \code{x4}), and
a time variable (\code{time}) from \code{data}.
All extracted variables are time series vectors.
}

