% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impl_cond.r
\name{change.nested.call}
\alias{change.nested.call}
\title{recursively change element}
\usage{
change.nested.call(x, adapt.atomic = NULL, adapt.name = NULL,
  adapt.call = NULL, pre.adapt.call = NULL, adapt.pairlist = NULL, ...)
}
\description{
recursively change element
}

