% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cesModel.R
\name{extract.cesModel}
\alias{extract.cesModel}
\title{Extract information from cesModel objects for texreg}
\usage{
extract.cesModel(model)
}
\arguments{
\item{model}{a \code{cesModel} object from which data is to be extracted.}
}
\value{
a \code{texreg} object ready for further processing.
}
\description{
This function extracts information from a cesModel for texreg.
}
\examples{
if (require("EconData") & require("dplyr")) {
  extract(cesModel(iGDP ~ iK + iL + iQp + iYear, data = filter(Calvin, Country=="US"), nest = c(1,2,3)))
}
}

