% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.LINEXmodel}
\alias{fortify.LINEXmodel}
\title{fortify LINEX models}
\usage{
\method{fortify}{LINEXmodel}(model, data, ...)
}
\arguments{
\item{model}{a LINEX model}

\item{data}{a data frame.  Typically this is not needed
since the data are extra from \code{model}.  But this allows
a drop in replacement data set, if desired.}

\item{...}{additional arguments (not currently used)}
}
\description{
fortify LINEX models
}

