% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{getData}
\alias{getData}
\title{Data used to fit a model}
\usage{
getData(model)
}
\arguments{
\item{model}{the model object from which you want to extract the \code{data}.}
}
\value{
the data used to fit the model.
}
\description{
An convenience function that returns the data that was used to fit a model.
}

