% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_compute.r
\name{init.formula.partners}
\alias{init.formula.partners}
\title{Find partner variables for xcut and ycut formulas}
\usage{
init.formula.partners(em, fos)
}
\description{
Find partner variables for xcut and ycut formulas
}

