% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{linexModel}
\alias{linexModel}
\title{Fitting LINEX models}
\usage{
linexModel(formula, data, response, x1, x2, x3, time, save.data = TRUE)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ x1 + x2 + x3 + time}}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x3}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}
}
\description{
Fitting LINEX models
}

