% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.r
\name{make.shocks.table}
\alias{make.shocks.table}
\title{transform shocks list into a data frame with one row
for each (t, shock, par) combination}
\usage{
make.shocks.table(shocks)
}
\description{
transform shocks list into a data frame with one row
for each (t, shock, par) combination
}

