% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{natmetaFrame}
\alias{natmetaFrame}
\title{Natural coefficients and metadata for all model attempts}
\usage{
natmetaFrame(object)
}
\arguments{
\item{object}{the model object from which you want to extract
information from all model attempts.}
}
\value{
a data frame containing the \code{naturalCoeffs} and \code{meta} attributes from
all model attempts in \code{object}.
}
\description{
A convenience function that returns a data frame containing both
\code{naturalCoeffs} and \code{meta} attributes for all model attempts
stored in the \code{model.attempts} attribute of \code{object}.
}

