% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{naturalCoef}
\alias{naturalCoef}
\title{Natural coefficients of a model}
\usage{
naturalCoef(object)
}
\arguments{
\item{object}{the model object from which you want to extract the \code{meta} attribute.}
}
\value{
the \code{naturalCoeffs} attribute from \code{object}.
}
\description{
A convenience function that returns the \code{naturalCoeffs} attribute of a model object.
}

