% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plm.R
\name{plm}
\alias{plm}
\title{Parameterized Linear Models}
\usage{
plm(formula, data = parent.frame(), params = c(), optimize = TRUE,
  .ocall = NULL, method = "nlm", ..., debug = FALSE)
}
\arguments{
\item{formula}{a formula describing the model}

\item{data}{a data frame}

\item{params}{a named vector of the parameters to be fit by numerical optimation outside of \code{lm}.}

\item{optimize}{a logical indicating whether \code{param} should be plugged in or used as the seed
for the numerical optimization}

\item{.ocall}{used for recursive calling}

\item{...}{additional arguments, currently ignored.}
}
\value{
On object of class \code{c("plm", "lm")} which is an enhanced \code{"lm"} object.
}
\description{
The optimal model is fit in a case where asside from some parameters, the model is linear.
Numerical optimization is used over these parameters with \code{lm} fitting the rest.
}
\examples{
if( require(EconData) & require(dplyr) ) {
plm( log(iGDP) - log(iK) ~ iYear,
     data=Calvin \%>\% filter(Country=="US"))

plm( log(iGDP) - delta * iK - (1-delta) * iL ~ iYear,
     data=Calvin \%>\% filter(Country=="US"),
     p=c(delta=0.4),
     optimize=FALSE )

plm( log(iGDP) - delta * iK - (1-delta) * iL ~ iYear, data=Calvin \%>\% filter(Country=="US"),
     p=c(delta=0.4),
     optimize=TRUE )

plm( log(iGDP) - delta * iK - (1-delta) * iL ~ iYear, data=Calvin \%>\% filter(Country=="US"),
     p=c(delta=0.4), method=c("nlm", "spg"),
     optimize=TRUE )

foo <- plm( log(iGDP / delta*(delta_1*iK + (1-delta_1)*iL) + (1-delta)*iQp) ~ iYear,
            data=Calvin \%>\% filter(Country=="US"),
            params=c(delta=0.5, delta_1=0.5))
foo
class(foo)
}
}

