% Generated by roxygen2 (4.0.2): do not edit by hand
\name{response}
\alias{response}
\title{Return response values from original data used to fit a model}
\usage{
response(object, ...)
}
\arguments{
\item{object}{a model object from one of the model fitting functions.}

\item{...}{additional arguments}
}
\value{
a numeric vector
}
\description{
This function returns the values of the original response variable
The values are calculated from fits and residuals.
For models of class \code{"LINEXmodel"}
\code{"CDEmodel"} or \code{"cesEst"}, the logarthmic transformation, if it
was used, will be undone, returning the values to their natural scale.
}

