% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{rowsSameOrdered}
\alias{rowsSameOrdered}
\title{Tells whether variables are always similarly ordered across all observations.}
\usage{
rowsSameOrdered(data)
}
\arguments{
\item{data}{a data frame that contains rows that are to be compared for ordering}
}
\value{
\code{TRUE} if all rows are ordered the same, \code{FALSE} if they are not.
}
\description{
Tells whether variables are always similarly ordered across all observations.
}
\note{
If all rows are comprised of equal numbers, \code{FALSE} is returned.
}

