% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{sfModel}
\alias{sfModel}
\title{Fitting single factor models}
\usage{
sfModel(formula, data, response, factor, time, constrained = FALSE,
  save.data = TRUE)
}
\arguments{
\item{formula}{a formula of the form \code{ y ~ factor + time }}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{constrained}{a logical indicating whether the model parameters are constrained}
}
\value{
an lm object with some additional attributes
}
\description{
Fitting single factor models
}
\examples{
US <- subset(Calvin, Country=="US")
sfModel(response = iGDP, factor=iK, time = iYear, data=US)
sfModel(iGDP ~ iK + iYear, data=US)
}

