% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{sse}
\alias{sse}
\title{Model metadata}
\usage{
sse(object)
}
\arguments{
\item{object}{the model object from which you want to extract the \code{meta} attribute.}
}
\value{
the \code{meta} attribute from \code{object}.
}
\description{
A convenience function that returns the \code{meta} attribute of a model object.
}

