% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphics.R
\name{standardScatterPlot}
\alias{standardScatterPlot}
\title{#}
\usage{
standardScatterPlot(data, mapping = aes(), orig_color = "gray70",
  size = 2, alpha = 0.4, facet_formula = countryAbbrev ~ nest)
}
\description{
#
}

