% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Graphics-backup.R, R/Graphics.R
\name{tri2x}
\alias{tri2x}
\title{Some Graphics functions}
\usage{
tri2x(x, y, z)

tri2x(x, y, z)
}
\description{
Some graphics functions for working with Economic Models.

Some graphics functions for working with Economic Models.
}

