% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trianglePlots.R
\docType{data}
\name{StatTriangle}
\alias{StatTriangle}
\alias{stat_triangle}
\alias{tri_grid}
\alias{tri_labels}
\title{Triangle Plots}
\format{\preformatted{Classes 'StatTriangle', 'Stat', 'ggproto' <ggproto object: Class StatTriangle, Stat>
    compute_group: function
    compute_layer: function
    compute_panel: function
    default_aes: uneval
    non_missing_aes: 
    parameters: function
    required_aes: x y z
    retransform: TRUE
    setup_data: function
    setup_params: function
    super:  <ggproto object: Class Stat> 
}}
\usage{
StatTriangle

stat_triangle(mapping = NULL, data = NULL, geom = "point",
  position = "identity", show.legend = NA, inherit.aes = TRUE, ...)

tri_grid(colour = color, color = "gray70", grid_lines = 5,
  labels = c("alpha[1]", "alpha[2]", "alpha[3]"), ...)

tri_labels(labels = c("alpha[1]", "alpha[2]", "alpha[3]"), parse = TRUE,
  colour = color, color = "gray60", size = 4, padding = 0.1, ...)
}
\description{
Triangle Plots
}
\examples{
ddd <- data.frame( x=c(1, 0, .5, .25, 0), y = c(0, .2, .2, .25, .8)) \%>\% mutate(z = 1 - x - y)
ggplot()  +
  tri_theme() +
  tri_grid(colour="red") +
  tri_labels(size=10, padding=.05) +
  geom_path(data=ddd, aes(x=x, y=y, z=z), stat="triangle", size=2, alpha=.3) +
  geom_point(data=ddd, aes(x=x, y=y, z=z), stat="triangle", size=3, alpha=.6, colour="navy")
}
\keyword{datasets}

