% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesModel.R
\name{withinConstraints}
\alias{withinConstraints}
\title{Tells whether a CES model is within constraints}
\usage{
withinConstraints(model)
}
\arguments{
\item{model}{the CES model to be evaluated}
}
\value{
\code{TRUE} if the fitted model is within the constraints, \code{FALSE} if not.
}
\description{
This function evaluates a fitted CES model for economic meaningfullness as expressed in constraints on 
\code{delta_1}, \code{sigma_1} and, if present, \code{delta_2}, \code{sigma_2}, \code{delta}, and \code{sigma}.
Specifically, this function checks whether \code{0 <= delta <= 1} and \code{sigma >= 0}
}
\note{
Missing, \code{NULL}, and \code{NA} parameters are interpreted as meeting constraints.

If none of \code{delta_1}, \code{sigma_1}, \code{delta_2}, \code{sigma_2}, \code{delta}, and \code{sigma}
is present, \code{FALSE} is returned.

This function relies upon the presence of the \code{naturalCoeffs} attribute to model.
If \code{naturalCoeffs} is \code{NULL}, \code{FALSE} is returned.
}
