% Generated by roxygen2 (4.0.1): do not edit by hand
\name{yhat}
\alias{yhat}
\title{Compute fitted values on natural scale}
\usage{
yhat(object, ...)
}
\arguments{
\item{object}{An object returned by a fitting function.}
}
\description{
This is similar to \code{fitted}, but will invert logarithmic
transformations of the response variable for certain models (e.g., LINEX, and Cobb-Douglas
models fit in this package).
}

