% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{apply_lm}
\alias{apply_lm}
\title{Apply lm to (several) formula templates}
\usage{
apply_lm(formula, data, formulaTemplates, coefNames, save.data = TRUE, ...)
}
\arguments{
\item{formula}{A formula containing the (common) response on the left and all variables
used in any of the models on the right.}

\item{data}{a data frame containing the necessary variables.}

\item{formulaTemplates}{a list of formulas using response \code{y} and
predictors
\code{capital},
\code{labor},
\code{energy},
and
\code{time}.}

\item{coefNames}{a list of names for the coefficients of the fitted model.}

\item{save.data}{a logical, currently not used.}

\item{...}{additional arguments, currently not used.}
}
\description{
This automates fitting several models that make use of the same variables and data set.
}
\examples{
apply_lm( iGDP ~ iK + iL + iQp + iYear, data=Calvin,
  formulaTemplates = list(
    log(y) - log(capital) ~ time,
    log(y) - log(labor) ~ time,
    log(y) - log(energy) ~ time),
 coefNames = list(
   c("logscale", "lambda"),
   c("logscale", "lambda"),
   c("logscale", "lambda"))
 )
}

