% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{cd3Model}
\alias{cd3Model}
\title{Fitting Cobb-Douglas models with 3 Factors}
\usage{
cd3Model(formula, data, response, x1, x2, x3, time, constrained = FALSE,
  save.data = TRUE, ...)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ x1 + x2 + x3 + time}}

\item{data}{a data frame in which \code{formala} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x3}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{constrained}{a logical indicated whether the coefficents are constrained. See details}

\item{\dots}{additional arguments; currently unused.}
}
\description{
Fitting Cobb-Douglas models with 3 Factors
}
\details{
More about contranints TBA.
}
\examples{
US <- subset(Calvin, Country=="US")
cd3Model(response = iGDP, x1 = iK, x2 = iL, x3 = iQp, time = iYear, data=US)
cd3Model(iGDP ~ iK + iL + iQp + iYear, data=US)
cd3Model(response = "iGDP", x1="iK", x2="iL", x3="iQp", time="iYear", data=US)
}

