% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{cdModel}
\alias{cdModel}
\title{Fitting Cobb-Douglas models}
\usage{
cdModel(formula, data, response, x1, x2, x3, time, constrained = TRUE,
  save.data = TRUE, correlation = NULL, ...)
}
\arguments{
\item{data}{a data frame in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x3}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{constrained}{a logical indicating whether the parameters should be constrained in the fitting process.}

\item{formala}{a formula of the form 
\code{response ~ x1 + x2 + time} or 
\code{response ~ x1 + x2 + x3 + time}}
}
\value{
a CDEmodel object, which is an lm object with some additioanl attributes.
}
\description{
Fitting Cobb-Douglas models
}
\examples{
US <- subset(Calvin, Country=="US")
cdModel(response = "iGDP", x1 = "iK", x2 = "iL", time = "iYear", data=US)
cdModel(iGDP ~ iK + iL + iYear, data=US)
cdModel(response = "iGDP", x1 = "iK", x2 = "iL", x3 = "iQp", time = "iYear", data=US)
cdModel(iGDP ~ iK + iL + iQp + iYear, data=US)
cdModel(response="iGDP", x1="iK", x2="iL", x3="iQp", time="iYear", data=US)

}

