% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cdwoeModel}
\alias{cdwoeModel}
\title{Fitting Cobb-Douglas Models}
\usage{
cdwoeModel(formula, data, response, capital, labor, time, constrained = FALSE,
  save.data = TRUE, ...)
}
\arguments{
\item{formala}{a formual of the form \code{response ~ capital + labor + time}}

\item{data}{a data fram in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{capital}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{labor}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{constrained}{a logical indicating whether the parameters are constrained}
}
\value{
a CDEmodel object, which is an lm object with some additional attributes.
}
\description{
Fitting Cobb-Douglas Models
}

