% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{cesBoundaryModel}
\alias{cesBoundaryModel}
\title{Fits CES boundary models}
\usage{
cesBoundaryModel(data, f, nest, id)
}
\arguments{
\item{data}{historical time series data}

\item{f}{a CES formula in the form \code{y ~ a + b + c + d + time}}

\item{nest}{identifies the nesting of the variables in the original formula,
assumed to be of the form \code{c(1,2,3)}.}

\item{id}{the identification number (from Table 2) for the boundary model you want to fit.}
}
\value{
a model object with class \code{CESModel} with \code{naturalCoeffs} and \code{meta} attributes.
}
\description{
The boundary models are given in Table 2 of
Heun, et al "An Empirical Investigation of the Role of Energy in Economic Growth"
}
\note{
the \code{naturalCoeffs} attribute includes correct values of boundary parameters.
\code{NA} values indicate that the parameter is unknowable at that boundary.
}

