% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cesBoundaryModels.R
\name{cesBoundaryModels}
\alias{cesBoundaryModels}
\title{Fits CES boundary models}
\usage{
cesBoundaryModels(formula, data, nest, method = "nlm", subset = TRUE)
}
\arguments{
\item{formula}{a CES formula in the form \code{y ~ a + b + c + d + time}}

\item{data}{historical time series data}

\item{nest}{a permutation of the integers 1 to the number of factors describing how to nest and permute
the variables in the formula.}
}
\value{
a list of boundary models.
}
\description{
The boundary models are given in Table 2 of
Heun, et al "An Empirical Investigation of the Role of Energy in Economic Growth"
}
\examples{
if (require(EconData) & require(dplyr)) {
  cesBoundaryModels(iGDP ~ iK + iL + iQp + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(1,2,3))
  cesBoundaryModels(iGDP ~ iK + iL + iQp + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(2,3,1))
  cesBoundaryModels(iGDP ~ iK + iL + iYear, data=Calvin \%>\% filter(Country=="US"), nest=c(1,2))
  }
}

