% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fortify.LINEXmodel}
\alias{fortify.LINEXmodel}
\title{fortify LINEX models}
\usage{
fortify.LINEXmodel(model, data, ...)
}
\arguments{
\item{model}{a LINEX model}

\item{data}{a data frame.  Typically this is not needed
since the data are extra from \code{model}.  But this allows
a drop in replacement data set, if desired.}

\item{...}{additional arguments (not currently used)}
}
\description{
fortify LINEX models
}

