% Generated by roxygen2 (4.0.1): do not edit by hand
\name{historicalPlot}
\alias{historicalPlot}
\title{Plot historical data}
\usage{
historicalPlot(data, mapping, facet_formula, line_types)

historicalPlot(data, mapping, facet_formula, line_types)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame}

\item{data}{a data frame containing data to be plotted.}

\item{mapping}{the argument to the geom_line function.
Use for setting, x, y, color, among other things: mapping=aes(x=scale, y=lambda, color=country)}

\item{facet_formula}{a formula of the form \code{y ~ x} where \code{y} is the varible to facet in
the y direction
and \code{x} is the variable to facet in the x direction.}

\item{line_types}{is a vector of linetype identifiers to be applied as the line types for the \code{group}s.}
}
\description{
Plot historical data

Create graph of historical data.
}
\details{
This function returns a figure with facets specified by \code{facet_formula}
and various lines specified by \code{line_formula}.
You may have to use the package \code{reshape2} to "melt" your data to the correct form before passing
it to this function.
}

