% Generated by roxygen2 (4.0.1): do not edit by hand
\name{historicalPlot}
\alias{historicalPlot}
\title{Create graph of historical data.}
\usage{
historicalPlot(data, facet_formula, lines_formula, line_types)
}
\arguments{
\item{data}{a data frame containing data to be plotted.}

\item{facet_formula}{a formula of the form \code{y ~ x} where \code{y} is the varible to facet in
the y direction
and \code{x} is the variable to facet in the x direction.}

\item{lines_formula}{a formula of the form \code{group + y ~ x}
where \code{group} is the column of the data frame that identifies data to be plotted as discrete lines
(often the "variable" column in a melted data frame),
\code{y} is the column of the data frame that contains values to be plotted on the y axis
(often the "value" column in a melted data frame), and
\code{x} is the column of the data frame that contains values to be plotted on the x axis.}

\item{line_types}{is a vector of linetype identifiers to be applied as the line types for the \code{group}s.}
}
\description{
Create graph of historical data.
}
\details{
This function returns a figure with facets specified by \code{facet_formula}
and various lines as specified by \code{line_formula}.
You may have to use the package \code{reshape2} to "melt" your data to the correct form before passing
it to this function..
}

