% Generated by roxygen2 (4.0.1): do not edit by hand
\name{historicalPlot}
\alias{historicalPlot}
\title{Create graph of historical data.}
\usage{
historicalPlot(data, facet_formula, lines_formula, line_types)
}
\arguments{
\item{data}{a data frame containing the the historical data to be plotted.}

\item{data}{a data frame containing historical data.}

\item{facet_formula}{a formula of the form \code{y ~ x} where \code{y} is the varible to facet in
the y direction
and \code{x} is the variable to facet in the x direction.}

\item{lines_formula}{a formula of the form \code{y_1 + y_2 + y_3 + ... ~ x}
where \code{y_i} are the variables to be plotted on the y axis of each small graph
and \code{x} is the variable to be plotted on the x axis of each small graph.}

\item{line_types}{is a vector of values to be applied as the line types for the \code{y_i} variables.}
}
\description{
Create graph of historical data.
}
\details{
The function returns a figure with the facets specified by \code{facet_formula}
and various lines as specified by \code{line_formula}.
}

