% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{linexModel}
\alias{linexModel}
\title{Fitting LINEX models}
\usage{
linexModel(formula, data, response, capital, labor, energy, time,
  save.data = TRUE)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ capital + labor + energy + time}}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{capital}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{labor}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{energy}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}
}
\description{
Fitting LINEX models
}

