% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myqdata}
\alias{myqdata}
\title{Quantiles from Data}
\usage{
myqdata(p, vals, data = parent.frame(), ...)
}
\arguments{
\item{p}{a vector of probabilities}

\item{vals}{a numeric vector or a name of a variable in \code{data}}

\item{data}{a data frame}

\item{...}{addtional arguments passed to \code{\link{quantile}}}
}
\value{
a vector of quantiles
}
\description{
A thin wrapper around \code{\link{quantile}}
}
\examples{
myqdata( .5, eruptions, data=faithful)
}

