% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{resampledFits}
\alias{resampledFits}
\title{Perform resample fits for a model}
\usage{
resampledFits(model, method = c("residual", "resample", "wild", "parametric",
  "debug"), n, save.data = FALSE, seed, id, include.original = TRUE,
  reindex = FALSE, mc.cores = parallel::detectCores(), ...)
}
\arguments{
\item{model}{a model returned from \code{linexModel}, \code{cdModel}, 
\code{cesModel}, or \code{sfModel}.  \code{model} must
have been fit with \code{save.data = TRUE}.}

\item{method}{one of \code{"resample"}, \code{"residual"}, \code{"wild"}, \code{"parametric"}, or \code{"debug"}}

\item{n}{the number of resamples you want to perform}

\item{save.data}{a logical indicating whether the data should be saved 
with each model.}

\item{seed}{a seed for \code{set.seed} to make results reproducible}

\item{id}{a character vector of length 1 used as an identifier that will be 
added to the output.  This is convenient
for marking data internally if you are doing resampling on many different models.}

\item{include.original}{a logical indicating whether the original model and 
coefficients should be included in the returned value.}

\item{reindex}{a boolean that indicates whether to reindex the resampled fits.}

\item{mc.cores}{number of cores to use in main loop.}

\item{...}{additional arguments passed to \code{fitfun}}
}
\value{
a list of length two containing a data frame of coefficients and a list of models
}
\description{
This function returns a list containing models and coefficients from the resample fits.
}

