% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{sfModel}
\alias{sfModel}
\title{Fitting single factor models}
\usage{
sfModel(formula, data, response, factor, time, constrained = FALSE,
  save.data = TRUE, correlation = NULL)
}
\arguments{
\item{formula}{a formula of the form \code{ y ~ factor + time }}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{constrained}{a logical indicating whether the model parameters are constrained}

\item{correlation}{an optional \code{\link{corStruct}} object describing the
within-group correlation structure. See the documentation of \code{\link{corClasses}}
for a description of the available \code{corStruct} classes.
If a grouping variable is to be used,
it must be specified in the \code{form} argument to the \code{corStruct} constructor.
Defaults to \code{NULL}, corresponding to uncorrelated errors.}
}
\value{
an lm object with some additional attributes
}
\description{
Fitting single factor models
}
\examples{
US <- subset(Calvin, Country=="US")
sfModel(response = iGDP, factor=iK, time = iYear, data=US)
sfModel(iGDP ~ iK + iYear, data=US)
sfModel(response="iGDP", factor="iK", time="iYear", data=US)
}

