% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fitting.R
\name{sse}
\alias{sse}
\title{Compute SSE from a model object}
\usage{
sse(object)
}
\arguments{
\item{object}{a model object}
}
\description{
Compute SSE from a model object and also check that certain constraints are met
by the coefficients.
}

