% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{sse}
\alias{sse}
\title{Compute SSE from a model object}
\usage{
sse(object)
}
\arguments{
\item{object}{a model object that inherits from \code{cesModel}.}
}
\value{
a data frame with with variables \code{sse}, \code{constrained}, \code{sse.contrained}.
}
\description{
Compute SSE from a model object and also check that certain constraints are met
by the coefficients.
}
